ThinkPalmAsign/requirements.txt
pytest
requests
pytest-asyncio
pika
prometheus_client
faker
pyyaml
python-dotenv
allure-pytest
pytest-html
loguru